smells & compile

+++ src/GmshToGLTF.cpp
@@ -91,21 +91,21 @@ int main(int argc, char *argv[]) {
                         double z = vertex.Z();
 
                         // To buffer
-                        Utils::floatToBuffer(x, buffer.data);
-                        Utils::floatToBuffer(y, buffer.data);
-                        Utils::floatToBuffer(z, buffer.data);
+                        Utils::floatToBuffer((float)x, buffer.data);
+                        Utils::floatToBuffer((float)y, buffer.data);
+                        Utils::floatToBuffer((float)z, buffer.data);
                       });
         model.buffers.push_back(buffer);
 
         // Buffer views
-        bufferViewIndices.buffer = model.buffers.size() - 1;
+        bufferViewIndices.buffer = (int)model.buffers.size() - 1;
         bufferViewIndices.byteOffset = 0;
         bufferViewIndices.byteLength =
             surface.triangles.size() * 3 * __SIZEOF_INT__;
         bufferViewIndices.target = TINYGLTF_TARGET_ELEMENT_ARRAY_BUFFER;
         model.bufferViews.push_back(bufferViewIndices);
 
-        bufferViewVertices.buffer = model.buffers.size() - 1;
+        bufferViewVertices.buffer = (int)model.buffers.size() - 1;
         bufferViewVertices.byteOffset =
             surface.triangles.size() * 3 * __SIZEOF_INT__ + paddingLength;
         bufferViewVertices.byteLength =
@@ -114,7 +114,7 @@ int main(int argc, char *argv[]) {
         model.bufferViews.push_back(bufferViewVertices);
 
         // Accessors
-        accessorIndices.bufferView = model.bufferViews.size() - 2;
+        accessorIndices.bufferView = (int)model.bufferViews.size() - 2;
         accessorIndices.byteOffset = 0;
         accessorIndices.componentType = TINYGLTF_COMPONENT_TYPE_UNSIGNED_INT;
         accessorIndices.count = surface.triangles.size() * 3;
@@ -123,7 +123,7 @@ int main(int argc, char *argv[]) {
         accessorIndices.maxValues.push_back(surface.maxIndex);
         model.accessors.push_back(accessorIndices);
 
-        accessorVertices.bufferView = model.bufferViews.size() - 1;
+        accessorVertices.bufferView = (int)model.bufferViews.size() - 1;
         accessorVertices.byteOffset = 0;
         accessorVertices.componentType = TINYGLTF_COMPONENT_TYPE_FLOAT;
         accessorVertices.count = surface.vertices.size();
@@ -143,9 +143,9 @@ int main(int argc, char *argv[]) {
         model.materials.push_back(material);
 
         // Primitive
-        primitive.indices = model.accessors.size() - 2;
-        primitive.attributes["POSITION"] = model.accessors.size() - 1;
-        primitive.material = model.materials.size() - 1;
+        primitive.indices = (int)model.accessors.size() - 2;
+        primitive.attributes["POSITION"] = (int)model.accessors.size() - 1;
+        primitive.material = (int)model.materials.size() - 1;
         primitive.mode = TINYGLTF_MODE_TRIANGLES;
 
         // Mesh
@@ -167,11 +167,11 @@ int main(int argc, char *argv[]) {
                                {"b", tinygltf::Value(color.at(2))}})}}));
 
         // Node
-        node.mesh = model.meshes.size() - 1;
+        node.mesh = (int)model.meshes.size() - 1;
         model.nodes.push_back(node);
 
         // Scene
-        scene.nodes.push_back(model.nodes.size() - 1);
+        scene.nodes.push_back((int)model.nodes.size() - 1);
       });
 
   // Scene
